src/Flask/requirements.txt
geopy
numpy
pandas
flask
flask_cors
Pillow
requests
tensorflow
scikit-learn
